# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/articles/{articleId}" : {
      "get" : {
        "tags" : [ "article-controller" ],
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "article-controller" ],
        "operationId" : "update_1",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateArticleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "article-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/v1/articles/{articleId}/draft" : {
      "get" : {
        "tags" : [ "automation-controller" ],
        "operationId" : "draft",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleDraft"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "automation-controller" ],
        "operationId" : "saveDraft",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DraftRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleDraft"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "automation-controller" ],
        "operationId" : "deleteDraft",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/v1/articles/{articleId}/restore" : {
      "post" : {
        "tags" : [ "article-controller" ],
        "operationId" : "restore",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/articles/{articleId}/reopen" : {
      "post" : {
        "tags" : [ "article-controller" ],
        "operationId" : "reopen",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/articles/{articleId}/confirm" : {
      "post" : {
        "tags" : [ "article-controller" ],
        "operationId" : "confirm",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/monitoring/summary" : {
      "get" : {
        "tags" : [ "monitoring-controller" ],
        "operationId" : "summary",
        "parameters" : [ {
          "name" : "range",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MonitoringSnapshot"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/articles/{articleId}/versions" : {
      "get" : {
        "tags" : [ "article-controller" ],
        "operationId" : "versions",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ArticleVersionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/{articleId}/assets/{assetId}" : {
      "get" : {
        "tags" : [ "article-asset-controller" ],
        "operationId" : "content",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "assetId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
    "/articles/{articleId}/assets/{assetId}" : {
      "get" : {
        "tags" : [ "article-asset-controller" ],
        "operationId" : "content_1",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "assetId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UpdateArticleRequest" : {
        "type" : "object",
        "properties" : {
          "expectedVersion" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          },
          "title" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "summary" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          },
          "markdown" : {
            "type" : "string",
            "maxLength" : 100000,
            "minLength" : 0
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "maxLength" : 50,
              "minLength" : 0
            },
            "maxItems" : 15,
            "minItems" : 0
          },
          "keywords" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "maxLength" : 100,
              "minLength" : 0
            },
            "maxItems" : 30,
            "minItems" : 0
          },
          "titleEn" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "summaryEn" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          },
          "markdownEn" : {
            "type" : "string",
            "maxLength" : 100000,
            "minLength" : 0
          },
          "tagsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "maxLength" : 50,
              "minLength" : 0
            },
            "maxItems" : 15,
            "minItems" : 0
          },
          "keywordsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "maxLength" : 100,
              "minLength" : 0
            },
            "maxItems" : 30,
            "minItems" : 0
          }
        },
        "required" : [ "markdown", "summary", "title" ]
      },
      "ArticleResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "projectId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "sourceType" : {
            "type" : "string"
          },
          "sourceUrl" : {
            "type" : "string"
          },
          "sourceTitle" : {
            "type" : "string"
          },
          "sourceDescription" : {
            "type" : "string"
          },
          "targetAudience" : {
            "type" : "string"
          },
          "articleType" : {
            "type" : "string"
          },
          "knowledgeLevel" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "markdown" : {
            "type" : "string"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keywords" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "titleEn" : {
            "type" : "string"
          },
          "summaryEn" : {
            "type" : "string"
          },
          "markdownEn" : {
            "type" : "string"
          },
          "tagsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keywordsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "language" : {
            "type" : "string"
          },
          "sourceRevision" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "currentVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "DraftRequest" : {
        "type" : "object",
        "properties" : {
          "baseVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "markdown" : {
            "type" : "string"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keywords" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "titleEn" : {
            "type" : "string"
          },
          "summaryEn" : {
            "type" : "string"
          },
          "markdownEn" : {
            "type" : "string"
          },
          "tagsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keywordsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "ArticleDraft" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "tenantId" : {
            "type" : "string"
          },
          "articleId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "actorSubject" : {
            "type" : "string"
          },
          "baseVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "markdown" : {
            "type" : "string"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keywords" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "titleEn" : {
            "type" : "string"
          },
          "summaryEn" : {
            "type" : "string"
          },
          "markdownEn" : {
            "type" : "string"
          },
          "tagsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keywordsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ChannelPerformance" : {
        "type" : "object",
        "properties" : {
          "channelType" : {
            "type" : "string",
            "enum" : [ "DEV", "WORDPRESS", "DISCOURSE", "GITHUB_DISCUSSIONS", "X", "REDDIT", "HASHNODE", "MEDIUM", "MASTODON", "GHOST", "XIAOHONGSHU", "CSDN", "JUEJIN", "ZHIHU", "CNBLOGS", "SEGMENTFAULT", "V2EX", "OSCHINA", "LINKEDIN", "WECHAT_OFFICIAL", "JIANSHU", "TOUTIAO", "BILIBILI_COLUMN", "BLOG_51CTO", "TENCENT_CLOUD", "ALIBABA_CLOUD", "HUAWEI_CLOUD" ]
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "published" : {
            "type" : "integer",
            "format" : "int64"
          },
          "failed" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "MonitoringSnapshot" : {
        "type" : "object",
        "properties" : {
          "capturedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "windowStart" : {
            "type" : "string",
            "format" : "date-time"
          },
          "window" : {
            "type" : "string",
            "enum" : [ "HOURS_24", "DAYS_7", "DAYS_30" ]
          },
          "projectCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "projectActivityCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "projectsByStatus" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "articleCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "articleActivityCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "articlesByStatus" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "articlesBySource" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "jobCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "jobActivityCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "jobsByStatus" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "windowJobsByStatus" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "publicationCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "publicationActivityCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "publicationsByStatus" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "windowPublicationsByStatus" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "accountCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "accountsByStatus" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "coveredChannelCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "channelPerformance" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ChannelPerformance"
            }
          }
        }
      },
      "ArticleVersionResponse" : {
        "type" : "object",
        "properties" : {
          "articleId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "versionNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "markdown" : {
            "type" : "string"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keywords" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "titleEn" : {
            "type" : "string"
          },
          "summaryEn" : {
            "type" : "string"
          },
          "markdownEn" : {
            "type" : "string"
          },
          "tagsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keywordsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "createdBy" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      }
    }
  }
}